Include Key for Second Floor Balcony to Prevent Softlocks
Due to an assumption I had made, it was possible to become softlocked
within the Northern Undead Asylum. The assumption was that the primary
door leading to the Asylum Demon encounter was always open, but this is
not the case. Only after you've dropped into the arena from above did it
become possible to use the primary door again.

This means that the 2F East Key is always required, and if shuffled
should always be granted either in the earliest sphere possible.